Fix x64 version doesn't compile.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2834 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/ui/uiutil.cpp
@@ -374,7 +374,7 @@ unsigned int qm::UIUtil::getPreferredWidth(HWND hwnd,
 	ClientDeviceContext dc(hwnd);
 	ObjectSelector<HFONT> fontSelecter(dc, wnd.getFont());
 	SIZE size;
-	dc.getTextExtent(wstrText.get(), wcslen(wstrText.get()), &size);
+	dc.getTextExtent(wstrText.get(), static_cast<int>(wcslen(wstrText.get())), &size);
 	return size.cx + 2;
 }
